refactor(executors): 归位工具结果失败语义
