+++ src/executors/anthropic/claude/trace.ts
@@ -1,7 +1,7 @@
 import type { ToolCallInfo, TurnInfo } from '../../contracts/trace.js';
 import type { ClaudeMessage } from './protocol.js';
 import { safeSliceForJson } from '../../../shared/json-safe-truncation.js';
-import { isToolResultFailureText } from '../../../observability/text-signals.js';
+import { isToolResultFailureText } from '../../tool-call-status.js';
 import { normalizeToolIdentity } from '../../../shared/tool-identity.js';
 
 export function isClaudeResultMessage(message: ClaudeMessage): boolean {
